Update wasm typescript to latest serde derive Split color into own file
