+++ rgb-solver/src/solver/struct_defs.rs
@@ -4,25 +4,7 @@ use wasm_typescript_definition::TypescriptDefinition;
 use super::van::Van;
 use crate::solver::grid_state::GridState;
 
-
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize, TypescriptDefinition, Default, Hash)]
-pub struct Color {
-    pub label: String,
-    pub red: u8,
-    pub green: u8,
-    pub blue: u8,
-
-    pub color_index: ColorIndex,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Serialize, Deserialize, TypescriptDefinition, Default, Hash)]
-pub struct ColorIndex(pub usize);
-
-impl ColorIndex {
-    pub fn is_white(&self) -> bool {
-        return self.0 == 0
-    }
-}
+include!("structs/color.rs");
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Serialize, Deserialize, TypescriptDefinition, Default, Hash)]
 pub struct CellIndex(pub usize);
