+++ rgb-solver/src/solver/structs/color.rs
@@ -1,0 +1,20 @@
+
+
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize, TypescriptDefinition, Default, Hash)]
+pub struct Color {
+    pub label: String,
+    pub red: u8,
+    pub green: u8,
+    pub blue: u8,
+
+    pub color_index: ColorIndex,
+}
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Serialize, Deserialize, TypescriptDefinition, Default, Hash)]
+pub struct ColorIndex(pub usize);
+
+impl ColorIndex {
+    pub fn is_white(&self) -> bool {
+        return self.0 == 0
+    }
+}
