+++ wasm-typescript-definition/Cargo.toml
@@ -11,9 +11,10 @@ repository = "https://github.com/tcr/wasm-typescript-definition"
 proc-macro = true
 
 [dependencies]
-quote = "0.5"
-serde_derive_internals = "0.23"
-syn = "=0.13"
+quote = "0.6"
+serde_derive_internals = "0.24.1"
+syn = "=0.15"
+proc-macro2 = "0.4"
 serde = "1"
 wasm-bindgen = "0.2"
